+++ rich/__main__.py
@@ -1,11 +1,16 @@
+import colorsys
 import io
 from time import process_time
 
+from rich.color import Color
+from rich.columns import Columns
 from rich.console import Console, ConsoleOptions, RenderGroup, RenderResult
 from rich.markdown import Markdown
 from rich.measure import Measurement
 from rich.padding import Padding
 from rich.panel import Panel
+from rich.pretty import Pretty
+from rich.segment import Segment
 from rich.style import Style
 from rich.table import Table
 from rich.syntax import Syntax
@@ -14,28 +19,29 @@ from rich import box
 
 
 class ColorBox:
-    def __init__(self, start: int = 16):
-        self.start = start
-
     def __rich_console__(
         self, console: Console, options: ConsoleOptions
     ) -> RenderResult:
-        start = self.start
-        for color_start in range(start, start + 36, 6):
-            text = Text()
-            for color_no in range(color_start, color_start + 6):
-                text.append("  ", Style(bgcolor=f"color({color_no})"))
-            yield text
+        for y in range(0, 5):
+            for x in range(options.max_width):
+                h = x / options.max_width
+                l = 0.1 + ((y / 5) * 0.7)
+                r1, g1, b1 = colorsys.hls_to_rgb(h, l, 1.0)
+                r2, g2, b2 = colorsys.hls_to_rgb(h, l + 0.7 / 10, 1.0)
+                bgcolor = Color.from_rgb(r1 * 255, g1 * 255, b1 * 255)
+                color = Color.from_rgb(r2 * 255, g2 * 255, b2 * 255)
+                yield Segment("▄", Style(color=color, bgcolor=bgcolor))
+            yield Segment.line()
 
     def __rich_measure__(self, console: "Console", max_width: int) -> Measurement:
-        return Measurement(12, 12)
+        return Measurement(1, max_width)
 
 
 def make_test_card() -> Table:
     """Get a renderable that demonstrates a number of features."""
     table = Table.grid(padding=1, pad_edge=True)
     table.title = "Rich features"
-    table.add_column("Feature", no_wrap=True, justify="right", style="bold red")
+    table.add_column("Feature", no_wrap=True, justify="center", style="bold red")
     table.add_column("Demonstration")
 
     color_table = Table(
@@ -44,25 +50,27 @@ def make_test_card() -> Table:
         show_header=False,
         show_edge=False,
         pad_edge=False,
-        padding=0,
     )
-
-    color_table.add_row(*(ColorBox(16 + color * 36) for color in range(6)))
-
-    table.add_row(
-        "Colors",
-        RenderGroup(
-            "[bold yellow]256[/] colors or [bold green]16.7 million[/] colors [blue](if supported by your terminal)[/].",
-            Padding(color_table, (1, 0, 0, 0)),
+    color_table.add_row(
+        # "[bold yellow]256[/] colors or [bold green]16.7 million[/] colors [blue](if supported by your terminal)[/].",
+        (
+            "✓ [bold green]4-bit color[/]\n"
+            "✓ [bold blue]8-bit color[/]\n"
+            "✓ [bold magenta]Truecolor (16.7 million)[/]\n"
+            "✓ [bold yellow]Dumb terminals[/]\n"
+            "✓ [bold cyan]Automatic color conversion"
         ),
+        ColorBox(),
     )
+
+    table.add_row("Colors", color_table)
 
     table.add_row(
         "Styles",
         "All ansi styles: [bold]bold[/], [dim]dim[/], [italic]italic[/italic], [underline]underline[/], [strike]strikethrough[/], [reverse]reverse[/], and even [blink]blink[/].",
     )
 
-    lorem = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque in metus sed sapien ultricies pretium a at justo. Maecenas luctus velit et auctor maximus. Donec faucibus vel arcu id pretium."
+    lorem = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque in metus sed sapien ultricies pretium a at justo. Maecenas luctus velit et auctor maximus."
     lorem_table = Table.grid(padding=1, collapse_padding=True)
     lorem_table.pad_edge = False
     lorem_table.add_row(
@@ -89,33 +97,37 @@ def make_test_card() -> Table:
         return table
 
     table.add_row(
-        "CJK support",
-        Panel(
-            "该库支持中文，日文和韩文文本！",
-            expand=False,
-            border_style="red",
-            box=box.DOUBLE_EDGE,
-        ),
+        "Asian\nlanguage\nsupport",
+        ":flag_for_china:  该库支持中文，日文和韩文文本！\n:flag_for_japan:  ライブラリは中国語、日本語、韓国語のテキストをサポートしています\n:flag_for_south_korea:  이 라이브러리는 중국어, 일본어 및 한국어 텍스트를 지원합니다",
     )
 
-    emoji_example = (
-        "Render emoji code: :+1: :apple: :ant: :bear: :baguette_bread: :bus: "
+    markup_example = (
+        "[bold magenta]Rich[/] supports a simple [i]bbcode[/i] like [b]markup[/b] for [yellow]color[/], [underline]style[/], and emoji! "
+        ":+1: :apple: :ant: :bear: :baguette_bread: :bus: "
     )
-    table.add_row("Emoji", comparison(Text(emoji_example), emoji_example))
-
-    markup_example = "[bold magenta]Rich[/] supports a simple [i]bbcode[/i] like [b]markup[/b], you can use to insert [yellow]color[/] and [underline]style[/]."
-    table.add_row(
-        "Console markup",
-        comparison(Text(markup_example), markup_example),
-    )
+    table.add_row("Markup", markup_example)
 
     example_table = Table(
-        title="Star Wars box office", show_header=True, header_style="bold magenta"
+        show_edge=False,
+        show_header=True,
+        expand=False,
+        row_styles=["none", "dim"],
+        box=box.SIMPLE,
     )
-    example_table.add_column("Date", style="dim", no_wrap=True)
-    example_table.add_column("Title")
-    example_table.add_column("Production Budget", justify="right", no_wrap=True)
-    example_table.add_column("Box Office", justify="right", no_wrap=True)
+    example_table.add_column("[green]Date", style="green", no_wrap=True)
+    example_table.add_column("[blue]Title", style="blue")
+    example_table.add_column(
+        "[cyan]Production Budget",
+        style="cyan",
+        justify="right",
+        no_wrap=True,
+    )
+    example_table.add_column(
+        "[magenta]Box Office",
+        style="magenta",
+        justify="right",
+        no_wrap=True,
+    )
     example_table.add_row(
         "Dec 20, 2019",
         "Star Wars: The Rise of Skywalker",
@@ -124,7 +136,7 @@ def make_test_card() -> Table:
     )
     example_table.add_row(
         "May 25, 2018",
-        "[red]Solo[/red]: A Star Wars Story",
+        "[b]Solo[/]: A Star Wars Story",
         "$275,000,000",
         "$393,151,347",
     )
@@ -133,6 +145,12 @@ def make_test_card() -> Table:
         "Star Wars Ep. VIII: The Last Jedi",
         "$262,000,000",
         "[bold]$1,332,539,889[/bold]",
+    )
+    example_table.add_row(
+        "May 19, 1999",
+        "Star Wars Ep. [b]I[/b]: [i]The phantom Menace",
+        "$115,000,000",
+        "$1,027,044,677",
     )
 
     table.add_row("Tables", example_table)
@@ -150,7 +168,24 @@ def iter_last(values: Iterable[T]) -> Iterable[Tuple[bool, T]]:
         previous_value = value
     yield True, previous_value'''
 
-    table.add_row("Syntax highlighting", Syntax(code, "python3", line_numbers=True))
+    pretty_data = {
+        "foo": [
+            3.1427,
+            (
+                "Paul Atriedies",
+                "Vladimir Harkonnen",
+                "Thufir Haway",
+            ),
+        ],
+        "atomic": (False, True, None),
+    }
+    table.add_row(
+        "Syntax\nhighlighting\n&\npretty\nprinting",
+        comparison(
+            Syntax(code, "python3", line_numbers=True, indent_guides=True),
+            Pretty(pretty_data, indent_guides=True),
+        ),
+    )
 
     markdown_example = """\
 # Markdown
@@ -162,10 +197,13 @@ Supports much of the *markdown*, __syntax__!
 - Block quotes
 - Lists, and more...
     """
-    table.add_row("Markdown", comparison(markdown_example, Markdown(markdown_example)))
+    table.add_row(
+        "Markdown", comparison("[cyan]" + markdown_example, Markdown(markdown_example))
+    )
 
     table.add_row(
-        "And more", """Progress bars, styled logging handler, tracebacks, etc..."""
+        "+more!",
+        """Progress bars, columns, styled logging handler, tracebacks, etc...""",
     )
     return table