+++ rich/pretty.py
@@ -19,7 +19,8 @@ from typing import (
 
 from rich.highlighter import ReprHighlighter
 
-from .__init__ import get_console
+from .abc import RichRenderable
+from . import get_console
 from ._pick import pick_bool
 from .cells import cell_len
 from .highlighter import ReprHighlighter
@@ -43,6 +44,8 @@ def install(
     crop: bool = False,
     indent_guides: bool = False,
     max_length: int = None,
+    max_string: int = None,
+    expand_all: bool = False,
 ) -> None:
     """Install automatic pretty printing in the Python REPL.
 
@@ -53,6 +56,8 @@ def install(
         indent_guides (bool, optional): Enable indentation guides. Defaults to False.
         max_length (int, optional): Maximum length of containers before abbreviating, or None for no abbreviation.
             Defaults to None.
+        max_string (int, optional): Maximum length of string before truncating, or None to disable. Defaults to None.
+        expand_all (bool, optional): Expand all containers. Defaults to False
     """
     from rich import get_console
 
@@ -65,12 +70,14 @@ def install(
             builtins._ = None  # type: ignore
             console.print(
                 value
-                if hasattr(value, "__rich_console__") or hasattr(value, "__rich__")
+                if isinstance(value, RichRenderable)
                 else Pretty(
                     value,
                     overflow=overflow,
                     indent_guides=indent_guides,
                     max_length=max_length,
+                    max_string=max_string,
+                    expand_all=expand_all,
                 ),
                 crop=crop,
             )
@@ -93,6 +100,7 @@ class Pretty:
         max_length (int, optional): Maximum length of containers before abbreviating, or None for no abbreviation.
             Defaults to None.
         max_string (int, optional): Maximum length of string before truncating, or None to disable. Defaults to None.
+        expand_all (bool, optional): Expand all containers. Defaults to False.
     """
 
     def __init__(
@@ -107,6 +115,7 @@ class Pretty:
         indent_guides: bool = False,
         max_length: int = None,
         max_string: int = None,
+        expand_all: bool = False,
     ) -> None:
         self._object = _object
         self.highlighter = highlighter or ReprHighlighter()
@@ -117,6 +126,7 @@ class Pretty:
         self.indent_guides = indent_guides
         self.max_length = max_length
         self.max_string = max_string
+        self.expand_all = expand_all
 
     def __rich_console__(
         self, console: "Console", options: "ConsoleOptions"
@@ -127,6 +137,7 @@ class Pretty:
             indent_size=self.indent_size,
             max_length=self.max_length,
             max_string=self.max_string,
+            expand_all=self.expand_all,
         )
         pretty_text = Text(
             pretty_str,
@@ -144,7 +155,11 @@ class Pretty:
 
     def __rich_measure__(self, console: "Console", max_width: int) -> "Measurement":
         pretty_str = pretty_repr(
-            self._object, max_width=max_width, indent_size=self.indent_size
+            self._object,
+            max_width=max_width,
+            indent_size=self.indent_size,
+            max_length=self.max_length,
+            max_string=self.max_string,
         )
         text_width = max(cell_len(line) for line in pretty_str.splitlines())
         return Measurement(text_width, text_width)
@@ -190,6 +205,11 @@ class Node:
     last: bool = False
     is_tuple: bool = False
     children: Optional[List["Node"]] = None
+
+    @property
+    def separator(self) -> str:
+        """Get separator between items."""
+        return "" if self.last else ","
 
     def iter_tokens(self) -> Iterable[str]:
         """Generate tokens for this node."""
@@ -247,7 +267,7 @@ class Node:
         Returns:
             str: A repr string of the original object.
         """
-        lines = [_Line(node=self)]
+        lines = [_Line(node=self, is_root=True)]
         line_no = 0
         while line_no < len(lines):
             line = lines[line_no]
@@ -264,6 +284,7 @@ class Node:
 class _Line:
     """A line in repr output."""
 
+    is_root: bool = False
     node: Optional[Node] = None
     text: str = ""
     suffix: str = ""
@@ -296,18 +317,20 @@ class _Line:
         else:
             yield _Line(text=node.open_brace, whitespace=whitespace)
         child_whitespace = self.whitespace + " " * indent_size
+        tuple_of_one = node.is_tuple and len(node.children) == 1
         for child in node.children:
+            separator = "," if tuple_of_one else child.separator
             line = _Line(
                 node=child,
                 whitespace=child_whitespace,
-                suffix="" if child.last else ",",
+                suffix=separator,
             )
             yield line
 
         yield _Line(
             text=node.close_brace,
             whitespace=whitespace,
-            suffix="" if node.last else ",",
+            suffix="," if (tuple_of_one and not self.is_root) else node.separator,
         )
 
     def __str__(self) -> str:
@@ -409,9 +432,9 @@ def pretty_repr(
     *,
     max_width: int = 80,
     indent_size: int = 4,
-    expand_all: bool = False,
     max_length: int = None,
     max_string: int = None,
+    expand_all: bool = False,
 ) -> str:
     """Prettify repr string by expanding on to new lines to fit within a given width.
 
@@ -419,11 +442,11 @@ def pretty_repr(
         _object (Any): Object to repr.
         max_width (int, optional): Desired maximum width of repr string. Defaults to 80.
         indent_size (int, optional): Number of spaces to indent. Defaults to 4.
-        expand_all (bool, optional): Expand all containers regardless of available width. Defaults to False.
         max_length (int, optional): Maximum length of containers before abbreviating, or None for no abbreviation.
             Defaults to None.
         max_string (int, optional): Maximum length of string before truncating, or None to disable truncating.
             Defaults to None.
+        expand_all (bool, optional): Expand all containers regardless of available width. Defaults to False.
 
     Returns:
         str: A possibly multi-line representation of the object.
@@ -443,9 +466,10 @@ def pprint(
     _object: Any,
     *,
     console: "Console" = None,
+    indent_guides: bool = True,
     max_length: int = None,
     max_string: int = None,
-    indent_guides: bool = True,
+    expand_all: bool = False,
 ):
     """A convenience function for pretty printing.
 
@@ -456,6 +480,7 @@ def pprint(
             Defaults to None.
         max_string (int, optional): Maximum length of strings before truncating, or None to disable. Defaults to None.
         indent_guides (bool, optional): Enable indentation guides. Defaults to True.
+        expand_all (bool, optional): Expand all containers. Defaults to False.
     """
     _console = get_console() if console is None else console
     _console.print(
@@ -464,7 +489,10 @@ def pprint(
             max_length=max_length,
             max_string=max_string,
             indent_guides=indent_guides,
-        )
+            expand_all=expand_all,
+            overflow="ignore",
+        ),
+        soft_wrap=True,
     )